Fixed Engineer bugs, now it works in all user flows
Our main bug was in the controller, we were getting error 500 from the server because we had async in the controller Get() and GetById() and it was not needed in that case, please check diff in git to see the changes that were made. Also, in DTOs wrong dto type was used in Managers list in EngineerEditDTO. Also, in EngineerEdit page instead of option select menu,  a label select manager was used in the select dropdown for manager. In the EngineerRepo there was left a "test" Include we used that was not needed, this was removed too.